moved hash and fetch infet, now runs on port 3069 always, dont change
