+++ src/components/modules/fet.js
@@ -3,12 +3,16 @@ let fet = async (url, method, body) => {
     if (!url) {
       reject("Invalid URL");
     }
+    if (! (url.startsWith("local") || url.startsWith("http"))) url = `http://localhost:6969${url}`
     if (!method) method = "GET";
-    //if (!body) body = {};
+    if (typeof(body) === "object") body = JSON.stringify(body)
     fetch(url, {
       method,
       mode: "cors",
       credentials: "include",
+      headers: {
+        "Content-Type": "application/json",
+      },
       body,
     })
       .then((data) => data.json())
@@ -22,4 +26,21 @@ let fet = async (url, method, body) => {
   });
 };
 
-export { fet };
+let hash = async message => {
+  // encode as UTF-8
+  const msgBuffer = new TextEncoder().encode(message);
+
+  // hash the message
+  const hashBuffer = await crypto.subtle.digest("SHA-256", msgBuffer);
+
+  // convert ArrayBuffer to Array
+  const hashArray = Array.from(new Uint8Array(hashBuffer));
+
+  // convert bytes to hex string
+  const hashHex = hashArray
+    .map((b) => b.toString(16).padStart(2, "0"))
+    .join("");
+  return hashHex;
+}
+
+export { fet, hash };
