+++ src/http-common.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 export default axios.create({
-  baseURL: "http://localhost:8000",
+  baseURL: "http://localhost:6969",
   headers: {
     "Content-type": "application/json",
   },
